refactor(github): update import paths and clean up code
- Update import paths for RepoData and other types- Remove unused components and imports
- Add type definitions for ColorMap and ContributionTableProps
- Convert SummaryTab and ContributionTable to functional components
- Clean up unused code and comments
