+++ src/components/github/analysis/tabs/RepoTabs.tsx
@@ -1,5 +1,5 @@
 import { Box, Grow } from "@mui/material";
-import "@/../styles/FormStyles.css";
+import "@/styles/FormStyles.css";
 
 const RepoTabs = () => {
 	return (
