+++ src/routeTree.gen.ts
@@ -1,6 +1,6 @@
 /* prettier-ignore-start */
 
- 
+/* eslint-disable */
 
 // @ts-nocheck
